db lookup table and payment data type changed from double to decimal

+++ Repository/DTO/RequestDTO/BookingCreateRequest.cs
@@ -18,7 +18,7 @@ namespace Repository.DTO.RequestDTO
         public DateTime DropoffTime { get; set; }
         [Range(15, double.MaxValue)]
         public double bookingFee { get; set; } = 15;
-        public double carRentPrice { get; set; }
+        public decimal carRentPrice { get; set; }
         public int rentime { get; set; }
         public string rentType { get; set; } = "Days"; // hour, day, week, month
     }
