+++ Repository/Data/Entities/InvoiceItem.cs
@@ -13,8 +13,8 @@ namespace Repository.Data.Entities
         public string Item { get; set; }
         public string Description { get; set; }
         public int Quantity { get; set; }
-        public double UnitPrice { get; set; }
-        public double Total { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal Total { get; set; }
         public string? Note { get; set; }
 
         public Guid InvoiceId { get; set; }
